# lib,src: replace all C++ promises with JS promises