--- a/lib/fs.js
+++ b/lib/fs.js
@@ -213,10 +213,7 @@ function exists(path, callback) {
 
 Object.defineProperty(exists, internalUtil.promisify.custom, {
   value: (path) => {
-    const { createPromise, promiseResolve } = process.binding('util');
-    const promise = createPromise();
-    fs.exists(path, (exists) => promiseResolve(promise, exists));
-    return promise;
+    return new Promise((resolve) => fs.exists(path, resolve));
   }
 });