--- a/lib/internal/http2/core.js
+++ b/lib/internal/http2/core.js
@@ -113,7 +113,6 @@ const { isArrayBufferView } = require('internal/util/types');
 const { FileHandle } = process.binding('fs');
 const binding = process.binding('http2');
 const { ShutdownWrap } = process.binding('stream_wrap');
-const { createPromise, promiseResolve } = process.binding('util');
 const { UV_EOF } = process.binding('uv');
 
 const { StreamPipe } = internalBinding('stream_pipe');
@@ -2747,11 +2746,9 @@ function connect(authority, options, listener) {
 // Support util.promisify
 Object.defineProperty(connect, promisify.custom, {
   value: (authority, options) => {
-    const promise = createPromise();
-    const server = connect(authority,
-                           options,
-                           () => promiseResolve(promise, server));
-    return promise;
+    return new Promise((resolve) => {
+      const server = connect(authority, options, () => resolve(server));
+    });
   }
 });