--- a/lib/internal/util.js
+++ b/lib/internal/util.js
@@ -8,10 +8,7 @@ const {
 const { signals } = process.binding('constants').os;
 
 const {
-  createPromise,
   getHiddenValue,
-  promiseResolve,
-  promiseReject,
   setHiddenValue,
   arrow_message_private_symbol: kArrowMessagePrivateSymbolIndex,
   decorated_private_symbol: kDecoratedPrivateSymbolIndex
@@ -276,24 +273,21 @@ function promisify(original) {
   const argumentNames = original[kCustomPromisifyArgsSymbol];
 
   function fn(...args) {
-    const promise = createPromise();
-    try {
+    return new Promise((resolve, reject) => {
       original.call(this, ...args, (err, ...values) => {
         if (err) {
-          promiseReject(promise, err);
-        } else if (argumentNames !== undefined && values.length > 1) {
+          return reject(err);
+        }
+        if (argumentNames !== undefined && values.length > 1) {
           const obj = {};
           for (var i = 0; i < argumentNames.length; i++)
             obj[argumentNames[i]] = values[i];
-          promiseResolve(promise, obj);
+          resolve(obj);
         } else {
-          promiseResolve(promise, values[0]);
+          resolve(values[0]);
         }
       });
-    } catch (err) {
-      promiseReject(promise, err);
-    }
-    return promise;
+    });
   }
 
   Object.setPrototypeOf(fn, Object.getPrototypeOf(original));