--- a/src/node_util.cc
+++ b/src/node_util.cc
@@ -9,7 +9,6 @@ using v8::Context;
 using v8::FunctionCallbackInfo;
 using v8::Integer;
 using v8::Local;
-using v8::Maybe;
 using v8::Object;
 using v8::Private;
 using v8::Promise;
@@ -140,36 +139,6 @@ void WatchdogHasPendingSigint(const FunctionCallbackInfo<Value>& args) {
   args.GetReturnValue().Set(ret);
 }
 
-
-void CreatePromise(const FunctionCallbackInfo<Value>& args) {
-  Local<Context> context = args.GetIsolate()->GetCurrentContext();
-  auto maybe_resolver = Promise::Resolver::New(context);
-  if (!maybe_resolver.IsEmpty())
-    args.GetReturnValue().Set(maybe_resolver.ToLocalChecked());
-}
-
-
-void PromiseResolve(const FunctionCallbackInfo<Value>& args) {
-  Local<Context> context = args.GetIsolate()->GetCurrentContext();
-  Local<Value> promise = args[0];
-  CHECK(promise->IsPromise());
-  if (promise.As<Promise>()->State() != Promise::kPending) return;
-  Local<Promise::Resolver> resolver = promise.As<Promise::Resolver>();  // sic
-  Maybe<bool> ret = resolver->Resolve(context, args[1]);
-  args.GetReturnValue().Set(ret.FromMaybe(false));
-}
-
-
-void PromiseReject(const FunctionCallbackInfo<Value>& args) {
-  Local<Context> context = args.GetIsolate()->GetCurrentContext();
-  Local<Value> promise = args[0];
-  CHECK(promise->IsPromise());
-  if (promise.As<Promise>()->State() != Promise::kPending) return;
-  Local<Promise::Resolver> resolver = promise.As<Promise::Resolver>();  // sic
-  Maybe<bool> ret = resolver->Reject(context, args[1]);
-  args.GetReturnValue().Set(ret.FromMaybe(false));
-}
-
 void SafeGetenv(const FunctionCallbackInfo<Value>& args) {
   CHECK(args[0]->IsString());
   Utf8Value strenvtag(args.GetIsolate(), args[0]);
@@ -223,10 +192,6 @@ void Initialize(Local<Object> target,
   env->SetMethod(target, "stopSigintWatchdog", StopSigintWatchdog);
   env->SetMethodNoSideEffect(target, "watchdogHasPendingSigint",
                              WatchdogHasPendingSigint);
-
-  env->SetMethodNoSideEffect(target, "createPromise", CreatePromise);
-  env->SetMethod(target, "promiseResolve", PromiseResolve);
-  env->SetMethod(target, "promiseReject", PromiseReject);
 
   env->SetMethod(target, "safeGetenv", SafeGetenv);
 }