… test/parallel/test-promise-internal-creation.js