@BridgeAR 2018-05-19
@TimothyGu we discussed this in person (even though we have meeting notes) two weeks ago.