@TimothyGu 2018-05-19
> They also behave a tiny bit different than "regular" promises (they can not be monkey patched).

Isn't this a desirable property?