@TimothyGu 2018-05-19
> Monkey patching is for example still important for APMs

Async Hooks are also important to APMs, and monkey patching `Promise` would not allow `PROMISE` hooks from being called as expected. (By saying this I'm not saying not monkeypatching is better for APMs – only APM developers can say for sure. I am saying this should not be listed as one of the pros for allowing monkeypatching.)

> Not being able to monkey patch those goes AFAIK against the spec.

There is no spec for `util.promisify`, except the documentation and our previous behavior, the latter of which this PR is now breaking.