@devsnek 2018-05-19
a happy medium that can let changing these to be monkey-patchable in a separate pr can be using `SafePromise` from internal/safe_globals