@TimothyGu 2018-05-19
@devsnek We don't want to expose any `Safe*` objects to userland, as it would no longer be safe. Simply caching the `Promise` constructor at startup should suffice.