@benjamingr 2018-05-19
First - do we have benchmarks for this?

Second - monkey patching won’t work because of async functions anyway - the future for diagnostics is better hooks

Otherwise LGTM