@benjamingr 2018-05-19
@apapirovski 

> This will be somewhere around 10-30% slower than the current version. I tested this fairly recently in relation to promisify. I'll leave it up to others to decide whether that's acceptable.

A 30% slower `promisify` would beat half the point of having promisify itself in core. While I personally find it a useful utility anyway the speed benefit was a big part of the reason people were convinced to add it.