@BridgeAR 2018-05-19
@jasnell if this will land at all I would like to keep it in a single commit. I do not see a compelling reason for it to be split into smaller chunks.

@apapirovski do you still have the benchmark code you used? Any performance impact is heavily influenced by the task at hand. So benchmarking a promisified setTimeout should show the biggest impact if it has influence on the performance.

-----

A general thing: I feel we block to many things for reasons that will have little to no effect on real world applications.