@apapirovski 2018-05-19
> do you still have the benchmark code you used? Any performance impact is heavily influenced by the task at hand. So benchmarking a promisified setTimeout should show the biggest impact if it has influence on the performance.

I don't. That said, the biggest performance impact was on things that were the longest running. Stuff like timeout & immediates were actually more manageable (between 0 - 15%, can't remember what the exact criteria were for what was less impacted and what was more), but promisifying fs & similar showed the worst performance.