@BridgeAR 2018-05-19
@apapirovski I guess it depends on how that was tested. If they were "chained" as in executed one after the other, setTimeout will very likely not show any impact at all because the CPU overhead would be fine due to the timeout.

Anyhow, I am going to write a new benchmark.