@BridgeAR 2018-05-19
Thanks for the info anyways :+1: @apapirovski