@benjamingr 2018-05-19
> A general thing: I feel we block to many things for reasons that will have little to no effect on real world applications.

A main reason we have `promisify` is for performance and platform support - I want to be sure when we block that that we have a good reason why.

Quoting @stefpanner on an old @paulirish thread from Google Plus:

> "If the abstraction is as close to a zero cost abstraction, developers won't be tempted to short-circuit and absorb complexity themselves."

I'm very reluctant in making these changes if we can't justify them with gains. Just using `new Promise` instead of `createPromise` isn't a gain IMO if it sacrifices performance along the way.

I appreciate your effort here (really) - but I've had several people ask me if we can make `promisify` faster _at bluebird_ which already ships a faster `promisify` at the moment anyway. These performance concerns are very real in server applications - our server at Peer5 runs millions of concurrent promises based off promisified functions.