@jasnell 2018-05-19
I'd like to see the benchmarks also. I get that this is the right thing to do, but if the performance cost is too high currently, then it likely makes sense to wait a bit.

Would *still* prefer to see the commits separated out but that's non-blocking.