@benjamingr 2018-05-19
@jasnell 

> I get that this is the right thing to do

Other than potentially performance from inlining - why though? Especially since there is another solution (using promisify from v8-extras) that we know is fast and has the same inlining benefits.