@benjamingr 2018-05-23
@jasnell 

> Creating the Promise directly in JavaScript should be faster and is something that the JavaScript VM ought to be able to optimize.

As someone who's had to explain why JavaScript promises can be faster than the platform ones for years I get that part :D I'm just not sure we can justify the regression _right now_ as well as what V8 can do.

This is _definitely_ something I want to discuss at the session with @bmeurer at the collaborator summit.