@addaleax 2018-05-23
> The reason is that calling the C++ `CreatePromise` (and the corresponding `Resolve`/`Reject`) still involve a boundary cross into JS.

@jasnell Following the code paths in V8, I find it hard to verify that… can you explain? We do have the overhead of the call into C++, yes, but I don’t see any call back into JS from there…