@devsnek 2018-05-23
@addaleax i think jasnell means pushing the created promise onto the stack from c++