@BridgeAR 2018-05-25
Rebased.

Becnhmark after the rebase compared to 10.2.1:

```
                                           confidence improvement accuracy (*)   (**)  (***)
 util/promisify.js fn='dnsLookup' n=200000        ***      7.92 %       ±3.18% ±4.25% ±5.56%
 util/promisify.js fn='fsExists' n=200000         ***      4.75 %       ±1.09% ±1.45% ±1.90%
 util/promisify.js fn='fsStat' n=200000            **      2.87 %       ±1.74% ±2.33% ±3.05%
 util/promisify.js fn='plainFunction' n=10000000          -1.50 %       ±2.25% ±3.01% ±3.97%
```