@apapirovski 2018-05-25
@BridgeAR Could you include the benchmarks here? When I tested I had benchmarks that had improvements but then when I wrote ones that were more realistic to the actual use cases the performance plummeted. That said, it's possible it was done before the changes that landed in V8 6.6 (but those might be reverted soon due to the async/await situation?).