@benjamingr 2018-05-25
> Becnhmark after the rebase compared to 10.2.1:

Can you run this against V8's benchmarks from https://github.com/v8/promise-performance-tests