@apapirovski 2018-05-25
@benjamingr I might've misunderstood but I thought that's what the discussion was about in relation to the async hooks and async/await situation. But perhaps that's unrelated to this. I don't know how much that work was intertwined.