@benjamingr 2018-05-25
> There is indeed a change: bluebird becomes significantly faster in the parallel benchmark with this PR applied and takes a minor hit on the doxbee benchmark.

The interesting part isn't really the bluebird ones here since replacing `promisify` means that the promises it's getting back aren't bluebird - if you want to measure bluebird you can test vs. its own `Promise.promisify` (which is likely much faster but that's another story).

The results look good overall - the interesting ones are:

With this PR:

```js
Time(doxbee-async-es2017-native): 110.2 ms.
Time(doxbee-promises-es2015-native): 111.6 ms.
Time(fibonacci-async-es2017-native): 229.3 ms.
Time(parallel-async-es2017-native): 341 ms.
Time(parallel-promises-es2015-native): 347.6 ms.
```

Before this PR:

```js
Time(doxbee-async-es2017-native): 113 ms.
Time(doxbee-promises-es2015-native): 119.1 ms.
Time(fibonacci-async-es2017-native): 242.2 ms.
Time(parallel-async-es2017-native): 364.7 ms.
Time(parallel-promises-es2015-native): 349 ms.
```

It looks like there is no real regression between before this PR and after this PR. I'm pretty surprised by this since `promisify` having to allocate a closure was always a slow point.

Can you test one more version with `promisify` with `createPromise` from `v8-extras` like @bmeurer suggested above?